Fix ignored width parameter in str_rmd_wrap.
The width parameter was not passed down to internal functions
properly and documentation was not up to date. This is fixed in this
commit.

Internal functions were stripped of default arugments, as they
are anyways supposed only to be called by API functions. This avoids
setting the same default over and over again.

A test case was added to check whether the argument `width` is
properly handled by all functions.

Closes #12.